Refactor LogEmitter from Murmur into its own file in src/.
It's a natural refactoring and it'd be nice to be able to
use LogEmitter in Mumble as well.
