+++ src/LogEmitter.cpp
@@ -1,0 +1,15 @@
+// Copyright 2005-2016 The Mumble Developers. All rights reserved.
+// Use of this source code is governed by a BSD-style license
+// that can be found in the LICENSE file at the root of the
+// Mumble source tree or at <https://www.mumble.info/LICENSE>.
+
+#include "murmur_pch.h"
+
+#include "LogEmitter.h"
+
+LogEmitter::LogEmitter(QObject *p) : QObject(p) {
+};
+
+void LogEmitter::addLogEntry(const QString &msg) {
+	emit newLogEntry(msg);
+};
